flask_backend/requirements.txt
# Flask Core
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Authentication & Security
PyJWT==2.8.0
Authlib==1.2.1
requests==2.31.0
cryptography==41.0.7

# Database
psycopg2-binary==2.9.7
redis==5.0.1

# Async Tasks
celery==5.3.4

# Payment Processing
stripe==7.7.0

# AI Services
google-genai
pydantic==2.5.0

# LaTeX & File Processing
Pillow==10.1.0
PyMuPDF==1.23.8
reportlab # <--- ADD THIS LINE

# Validation & Serialization
marshmallow==3.20.1
webargs==8.3.0

# Environment & Configuration
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0
